feat: #81. 경험치 조회하는 api 설계

+++ backend/user/src/main/java/com/scf/user/profile/domain/entity/Character.java
@@ -24,7 +24,7 @@ public class Character {
     private Integer characterId;
 
     @Column(name = "exp")
-    private Long exp = 0L;
+    private Integer exp = 0;
 
     @Column(name = "character_type")
     private Integer characterType;
